requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.1
colorama==0.4.6
dj-database-url==3.1.2
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
fastapi==0.135.1
frozenlist==1.7.0
greenlet==3.3.2
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
magic-filter==1.0.12
multidict==6.6.4
packaging==26.1
pillow==12.2.0
propcache==0.3.2
psycopg2==2.9.11
psycopg2-binary==2.9.12
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.12.1
python-dotenv==1.2.2
python-telegram-bot==22.4
PyYAML==6.0.3
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.48
sqlparse==0.5.3
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uv==0.9.8
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
yarl==1.20.1
